+++ react/ProductLink.tsx
@@ -4,6 +4,7 @@ import { ModalContext } from 'vtex.modal-layout'
 import { useProduct } from 'vtex.product-context'
 import React, { useState, useEffect } from 'react'
 
+import hasChildren from './modules/hasChildren'
 import { AvailableContext } from './modules/mappings'
 import interpolateLink from './modules/interpolateLink'
 
@@ -13,12 +14,13 @@ interface Props {
   label: string
   href: string
   children: React.ReactNode
+  target?: string
 }
 
 const CSS_HANDLES = ['link', 'label', 'childrenContainer'] as const
 
-export default function ProductLink(props: Props) {
-  const { label, href, children } = props
+function ProductLink(props: Props) {
+  const { label, href, children, target } = props
   const productContext = useProduct()
   const handles = useCssHandles(CSS_HANDLES)
   const [prevHref, setPrevHref] = useState()
@@ -43,10 +45,23 @@ export default function ProductLink(props: Props) {
     setShouldReplaceUrl(Boolean(modalDispatch))
   }, [modalDispatch])
 
-  return (
+  return target && target !== '_self' ? (
+    <a target={target} href={resolvedLink} className={handles.link}>
+      {label && <span className={handles.label}>{label}</span>}
+      {hasChildren(children) && (
+        <div className={handles.childrenContainer}>{children}</div>
+      )}
+    </a>
+  ) : (
     <Link to={resolvedLink} className={handles.link} replace={shouldReplaceUrl}>
       {label && <span className={handles.label}>{label}</span>}
-      {children && <div className={handles.childrenContainer}>{children}</div>}
+      {hasChildren(children) && (
+        <div className={handles.childrenContainer}>{children}</div>
+      )}
     </Link>
   )
 }
+
+ProductLink.schema = { title: 'Product Link' }
+
+export default ProductLink