+++ react/StoreLink.tsx
@@ -3,17 +3,20 @@ import { Link } from 'vtex.render-runtime'
 import { useCssHandles } from 'vtex.css-handles'
 import { ModalContext } from 'vtex.modal-layout'
 
+import hasChildren from './modules/hasChildren'
+
 interface Props {
   label: string
   href: string
   children: React.ReactNode
+  target?: string
 }
 
 const { useModalDispatch } = ModalContext
 const CSS_HANDLES = ['link', 'label', 'childrenContainer']
 
-export default function StoreLink(props: Props) {
-  const { label, href, children } = props
+function StoreLink(props: Props) {
+  const { label, href, children, target } = props
   const handles = useCssHandles(CSS_HANDLES)
   const modalDispatch = useModalDispatch()
   const [shouldReplaceUrl, setShouldReplaceUrl] = useState(
@@ -24,10 +27,25 @@ export default function StoreLink(props: Props) {
     setShouldReplaceUrl(Boolean(modalDispatch))
   }, [modalDispatch])
 
-  return (
+  return target && target !== '_self' ? (
+    <a target={target} href={href} className={handles.link}>
+      {label && <span className={handles.label}>{label}</span>}
+      {hasChildren(children) && (
+        <div className={handles.childrenContainer}>{children}</div>
+      )}
+    </a>
+  ) : (
     <Link to={href} className={handles.link} replace={shouldReplaceUrl}>
       {label && <span className={handles.label}>{label}</span>}
-      {children && <div className={handles.childrenContainer}>{children}</div>}
+      {hasChildren(children) && (
+        <div className={handles.childrenContainer}>{children}</div>
+      )}
     </Link>
   )
 }
+
+StoreLink.schema = {
+  title: 'Link',
+}
+
+export default StoreLink