+++ CHANGELOG.md
@@ -7,12 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Schema to the blocks.
+- `target` prop to all links.
+
 ## [0.1.1] - 2020-02-13
 
 ### Fixed
-- Improper section for props was removed
+
+- Improper section for props was removed.
 
 ## [0.1.0] - 2020-01-31
 
 ### Added
+
 - Intial release.